implemented vulkan material system  - uses spirv-cross to access shader reflection data  - split between Pipeline and Material classes remove custom vulkan hpp header since sdk has been updated added update_and_integrate and populate_command_buffer methods to Scene update maths::perspective functions to use vulkan z 0:1 instead of opengl z -1:1 remove unused maths::hsv_to_rgb function, can re-add if I ever need it added equality/inequality operators for matrix types reverted get_bytes_from_file back to using std::byte instead of char changed get_bytes_from_file to use c stdio instead of fstreams  - will likely change other functions later for consistency improved ShaderModules constructor signatures changed vk_create_image_2D to also create an image view changed vulkan memory mapping to return std::byte* instead of void* fixed loading signed normalised textures added draw_submesh method to sq::Mesh fixed encoding of packed normals when loading sq::Mesh made sq::Mesh invert winding when swapYZ is set fix attribute types of mesh bones and weights allow freeing descriptor sets allocated with the window's pool added setup.hpp file to vkdemo updated vkdemo to use the material system enabled and fixed normal mapping generated a higher resolution dice normal map renamed VulkTestApp to DemoApp changed namespace from sqt to demo changed vkdemo to properly free descriptor sets on window resize and shutdown

+++ source/sqee/objects/Armature.cpp
@@ -392,7 +392,7 @@ Armature::Pose Armature::compute_pose_discrete(const Animation& animation, uint 
 
 void Armature::compute_extra_continuous(Vec4F& result, const Animation::Track& track, float time)
 {
-    const uint frameCount = track.track.size() / sizeof(Vec4F);
+    const uint frameCount = uint(track.track.size() / sizeof(Vec4F));
     const auto [factor, frameA, frameB] = impl_compute_blend_times(frameCount, time);
 
     impl_compute_blended_value(factor, frameA, frameB, track, result);
@@ -400,7 +400,7 @@ void Armature::compute_extra_continuous(Vec4F& result, const Animation::Track& t
 
 void Armature::compute_extra_discrete(Vec4F& result, const Animation::Track& track, uint time)
 {
-    const uint frameCount = track.track.size() / sizeof(Vec4F);
+    const uint frameCount = uint(track.track.size() / sizeof(Vec4F));
     SQASSERT(frameCount == 1u || time < frameCount, "discrete time out of range");
 
     impl_compute_discrete_value(time, track, result);
